feat: 컴포넌트 분리( 내 정보 모달창 close 안 됨)
