+++ src/components/MyPage/LoungeList.ts
@@ -3,15 +3,15 @@ type LoungeList = { id: number; name: string };
 const LoungeList = [
   {
     id: 1,
-    name: '멋사 명지대',
+    name: '6 0 2 1 2 4 0 4 하 나 둘 셋 넷 다 잘려라잘려라',
   },
   {
     id: 2,
-    name: '이 건 열 글 자 이 건 열 글 자',
+    name: '이 건 한 글 임 이 건 열 글 자 하 나',
   },
   {
     id: 3,
-    name: '할수있다할수있다있뮤',
+    name: '6 0 2 1 2 4 0 4 이 유 라 r e a c t11111',
   },
   {
     id: 4,
@@ -19,7 +19,7 @@ const LoungeList = [
   },
   {
     id: 5,
-    name: '다섯',
+    name: '2022 명지대 멋쟁이사자처럼 라운지',
   },
   // { 스크롤 구현 예정
   //   id: 6,
